Add missing install lock
There was a race condition between checking for folder conflicts and
extracting the archives.  A side effect is that archives are now
downloaded in advance (outside the lock) though they won't be used
in the case of a conflict.

This also simplifies the checking for folder conflicts.  Constructing
sets was unnecessary.