Make reconciliation less sucky
This fixes an issue with the folder
matcher matching add-ons with folders
that were not found in the `leftovers` set
which is to say it now performs a sub-set check
instead of an intersection.

The merge functions have been
consolidated into a single function
of mergacular goodness that acts more like
a merge-find set - whatever that is.

Last but not least you are now able to skip
items in the reconciliation prompt by pressing
<s>.  Yes, just like that.