Refactor several aspects of the GUI
Modified `resolve` so that it behaves like
other batch ops. Errors are captured in
resolve itself and won't be logged
repeatedly (once for each package) when
installing/updating/removing packages anymore.

Converted the response of RPC list methods
into a discriminated union-like object
(`AnyResponse`) to aid typing.

Simplified the structure of the add-on list
in the GUI so that there is always both a
reference add-on and a resolved add-on for
each add-on pair even when an add-on is not
installed or an installed add-on cannot be
resolved.

Replaced `search__searchStrategyExtra` with
`search__searchVersion` because the former
being an object would trigger spurious updates
in Svelte.