Create endpoint with random UUID
SOP and CORS do not apply to WebSockets.
We cannot validate the origin of the request
because a browser will open a socket concealing
the original origin.
WebSockets are very much intrinsically broken
in this regard.
For the time being we're appending a random UUID
to the WS endpoint which serves as an
impromptu access token.