Rework WA updater
This removes the `account` option from the WA updater.
instawow will now indiscriminately parse saved vars from
all accounts.
It also adds two-way caching of extracted
auras and Wago responses which was
a prerequisite for adding support in the GUI.
Finally, it adds the auto-update variant
of the Companion add-on to the catalogue
meaning that the Companion add-on will appear
in search results and will be reconcilable.

Addresses some of the concerns raised in #53.